Convert sunSearch to adapter interface
