+++ src/fswAlgorithms/attGuidance/sunSearch/sunSearch.cpp
@@ -1,7 +1,7 @@
 /*
  ISC License
 
- Copyright (c) 2024, Laboratory for Atmospheric and Space Physics, University of Colorado at Boulder
+ Copyright (c) 2025, Laboratory for Atmospheric and Space Physics, University of Colorado at Boulder
 
  Permission to use, copy, modify, and/or distribute this software for any
  purpose with or without fee is hereby granted, provided that the above
@@ -18,9 +18,6 @@
  */
 
 #include "sunSearch.h"
-#include "architecture/utilities/macroDefinitions.h"
-#include "architecture/utilities/avsEigenSupport.h"
-#include <cmath>
 
 /*! This method is used to reset the module.
  @return void
@@ -32,21 +29,8 @@ void SunSearch::reset(uint64_t currentSimNanos) {
     if (!this->vehConfigInMsg.isLinked()) {
         throw std::invalid_argument("SunSearch.vehConfigInMsg wasn't connected.");
     }
-    if (this->numberOfSlews != NUM_SLEWS) {
-        throw std::invalid_argument("The number of specified slew maneuvers must be equal to 3");
-    }
 
-    /*! read vehicle configuration message */
-    VehicleConfigMsgPayload vehConfigIn = this->vehConfigInMsg();
-    this->principleInertias[0] = vehConfigIn.ISCPntB_B[0];
-    this->principleInertias[1] = vehConfigIn.ISCPntB_B[4];
-    this->principleInertias[2] = vehConfigIn.ISCPntB_B[8];
-
-    for (int index = 0; index < NUM_SLEWS; index++) {
-        this->computeKinematicProperties(index);
-    }
-
-    this->resetTime = currentSimNanos;
+    this->algorithm.reset(currentSimNanos, this->vehConfigInMsg());
 }
 
 /*! This method is the main carrier for the computation of the guidance message
@@ -54,116 +38,18 @@ void SunSearch::reset(uint64_t currentSimNanos) {
  @param currentSimNanos The current simulation time for system
  */
 void SunSearch::updateState(uint64_t currentSimNanos) {
-    /*! create and zero the output message */
-    AttGuidMsgPayload attGuidOut{};
+    NavAttMsgPayload navAttIn = this->attNavInMsg();
+    AttGuidMsgPayload attGuidOut = this->algorithm.update(currentSimNanos, navAttIn);
 
-    /*! read vehicle configuration message */
-    NavAttMsgPayload attNavIn = this->attNavInMsg();
-
-    ReferenceMotionOutput referenceMotion{};
-
-    double CurrentSimSeconds = (currentSimNanos - this->resetTime) * NANO2SEC;
-
-    double timeInf = 0;
-    double timeSup = this->kinematicProperties[0].slewTotalTime;
-    for (int index = 0; index < NUM_SLEWS; ++index) {
-        if (CurrentSimSeconds >= timeInf && CurrentSimSeconds < timeSup) {
-            referenceMotion = this->computeReferenceMotion(currentSimNanos, index);
-            break;
-        } else if (CurrentSimSeconds >= timeSup && index != NUM_SLEWS - 1) {
-            timeInf += this->kinematicProperties[index].slewTotalTime;
-            timeSup += this->kinematicProperties[index + 1].slewTotalTime;
-        }
-    }
-
-    Eigen::Vector3d omega_BR_B = Eigen::Map<const Eigen::Vector3d>(attNavIn.omega_BN_B) - referenceMotion.omega_RN_B;
-
-    eigenVector3d2CArray(referenceMotion.omega_RN_B, attGuidOut.omega_RN_B);
-    eigenVector3d2CArray(omega_BR_B, attGuidOut.omega_BR_B);
-    eigenVector3d2CArray(referenceMotion.domega_RN_B, attGuidOut.domega_RN_B);
-
-    /*! Write the output messages */
     this->attGuidOutMsg.write(&attGuidOut, this->moduleID, currentSimNanos);
-}
-
-/*! Define this method to compute the kinematic properties of each slew
-    @return void
-    */
-void SunSearch::computeKinematicProperties(int const index) {
-    SlewProperties *SP = &this->slewProperties[index];
-    int axis = SP->slewRotAxis - 1;
-    double maxAcc = SP->slewMaxTorque / this->principleInertias[axis];
-
-    /*! Computing fastest bang-bang slew with no coasting arc */
-    double alpha = 4 * SP->slewAngle / (SP->slewTime * SP->slewTime);
-    double omegaMax = 2 * SP->slewAngle / SP->slewTime;
-    double totalTime = SP->slewTime;
-    double thrustTime = totalTime / 2;
-
-    /*! If angular acceleration exceeds limit, decrease acceleration and increase slew time */
-    if (alpha > maxAcc) {
-        alpha = maxAcc;
-        totalTime = 2 * sqrt(SP->slewAngle / alpha);
-        thrustTime = totalTime / 2;
-        omegaMax = alpha * thrustTime;
-    }
-
-    /*! If angular rate exceeds limit, increase slew time adding a coasting arc */
-    if (omegaMax > SP->slewMaxRate) {
-        omegaMax = SP->slewMaxRate;
-        totalTime = SP->slewAngle / omegaMax + omegaMax / alpha;
-        thrustTime = omegaMax / alpha;
-    }
-
-    KinematicProperties *KP = &this->kinematicProperties[index];
-
-    KP->slewRotAxis = SP->slewRotAxis;
-    KP->slewAngAcc = alpha;
-    KP->slewOmegaMax = omegaMax;
-    KP->slewTotalTime = totalTime;
-    KP->slewThrustTime = thrustTime;
-}
-
-/*! Define this method to compute the rate and acceleration as function of time
-    @return ReferenceMotionOutput
-    */
-ReferenceMotionOutput SunSearch::computeReferenceMotion(uint64_t const currentSimNanos, int const index) {
-    double zeroTime = 0;
-    for (int i = 0; i < index; ++i) {
-        zeroTime += this->kinematicProperties[i].slewTotalTime;
-    }
-    double localSimSeconds = (currentSimNanos - this->resetTime) * NANO2SEC - zeroTime;
-
-    KinematicProperties KP = this->kinematicProperties[index];
-    int axis = KP.slewRotAxis - 1;
-
-    Eigen::Vector3d omega_RN{Eigen::Vector3d::Zero()};
-    Eigen::Vector3d domega_RN{Eigen::Vector3d::Zero()};
-
-    if (localSimSeconds <= KP.slewThrustTime) {
-        omega_RN[axis] = KP.slewOmegaMax * localSimSeconds / KP.slewThrustTime;
-        domega_RN[axis] = KP.slewAngAcc;
-    } else if (localSimSeconds > KP.slewThrustTime && localSimSeconds < KP.slewTotalTime - KP.slewThrustTime) {
-        omega_RN[axis] = KP.slewOmegaMax;
-    } else if (localSimSeconds >= KP.slewTotalTime - KP.slewThrustTime && localSimSeconds <= KP.slewTotalTime) {
-        omega_RN[axis] = KP.slewOmegaMax * (KP.slewTotalTime - localSimSeconds) / KP.slewThrustTime;
-        domega_RN[axis] = -KP.slewAngAcc;
-    }
-
-    ReferenceMotionOutput referenceMotion{};
-    referenceMotion.omega_RN_B = omega_RN;
-    referenceMotion.domega_RN_B = domega_RN;
-
-    return referenceMotion;
 }
 
 /**
  * @brief Set the properties of a slew maneuver
  * @param slewPropertiesInput the properties of the slew maneuver
  */
-void SunSearch::setSlewProperties(SlewProperties slewPropertiesInput){
-    this->slewProperties[this->numberOfSlews] = slewPropertiesInput;
-    this->numberOfSlews += 1;
+void SunSearch::setSlewProperties(SlewProperties slewPropertiesInput) {
+    this->algorithm.setSlewProperties(slewPropertiesInput);
 }
 
 /**
@@ -172,7 +58,7 @@ void SunSearch::setSlewProperties(SlewProperties slewPropertiesInput){
  * @param index index of the slew maneuver
  */
 void SunSearch::modifySlewProperties(SlewProperties slewPropertiesInput, uint32_t index) {
-    this->slewProperties[index] = slewPropertiesInput;
+    this->algorithm.modifySlewProperties(slewPropertiesInput, index);
 }
 
 /**
@@ -180,6 +66,4 @@ void SunSearch::modifySlewProperties(SlewProperties slewPropertiesInput, uint32_
  * @param index index of the slew maneuver
  * @return SlewProperties the properties of the slew maneuver
  */
-SlewProperties SunSearch::getSlewProperties(uint32_t index) const {
-    return this->slewProperties[index];
-}
+SlewProperties SunSearch::getSlewProperties(uint32_t index) const { return this->algorithm.getSlewProperties(index); }
